feat: handle arguments and no results related errors
This commit includes a custom runtime exception for not found element as well as controller exception handler with tests.

+++ src/test/java/com/kairosdstest/prices/adapters/rest/PriceControllerTest.java
@@ -45,7 +45,7 @@ public class PriceControllerTest {
         assertEquals(expectedPriceList, response.getPriceList());
         assertEquals(expectedStartDate, response.getStartDate());
         assertEquals(expectedEndDate, response.getEndDate());
-        assertEquals(expectedFinalPrice, response.getPrice());
+        assertEquals(BigDecimal.valueOf(expectedFinalPrice), response.getPrice());
     }
 
     // Helper method to create a mocked Price object
